pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dynaconf>=3.2.10",
    "faker>=36.1.1",
    "fastapi[standard]>=0.115.11",
    "flask>=3.1.0",
    "instructor>=1.7.2",
    "ipykernel>=6.29.5",
    "langchain>=0.3.19",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "qdrant-client>=1.13.2",
    "requests>=2.32.3",
    "sentence-transformers>=3.4.1",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
]
